Prueba_DockerFileProducer/Productor/requirements.txt
schedule
pandas
kafka-python